Follow-up PR after upgrading to Helidon 4.1.5 (#9586)
* New type of connection pool using arrays.


* Removes print statement.


* Adds support for size-factor in config.


* Sets size-factor to 1.0.


* Removes old connection pool implementation that is no longer in use. Some other minor cleanup.

---------
